Store project and branch in the build table.
This commit adds initial support for storing the project and branch in
the build table.  This should make it possible to have more than one
Xenofarm project share the same database, which could be desirable if
you have related projects or branches.

server.pike support should be complete.  The result_parser should
continue to work.

The presentation modules have not been updated to support more than
one project in the same database.  As long as you only have one
project in the database, they should continue to work as before.

The new dbconvert.sql file converts the database, but you have to
change the "FIXME" string constants to name your project and branch.