Added the ability to ignore Git commits.
The new global variable ignored_globs defines a set of glob patterns.
If all the files changed in a commit are matched by at least one of
the globs, that commit is considered unimportant.  Unimportant commits
will not trigger a new build.  The code tries to handle forks and
merges in a sensible way.

New class: GitCommitNode.  New methods of the GitClient class:
commit_wanted(), first_wanted_commit().  Changed get_latest_checkin()
so that it returns first_wanted_commit() instead of current_commit().