Handle tasks per project.
Add a project column to the tasks table.  Update all users.

This is needed so that status pages for a build or branch don't need
to list tasks that are not relevant to them.  It also makes sense in
case two projects use the same task names, but run them in different
orders.

dbconvert.sql updates the database, but you need to replace FIXME with
the name of the project in the current database.