Added support for --web-format to server.pike.
Instead of storing files in the directory specified by --web-dir, the
directory be named using the --web-format pattern.  The following
formats are defined:

    %P: the project
    %B: the branch

If the pattern ends with "/", the basename of the build name is
appended, otherwise the pattern should contain the entire file name,
typically ending in .tar.gz.

Derived server scripts can add more patterns by overriding the
extra_web_formats() function.  The can also override the
build_stored() function which is called with the used file name once a
build is stored.