Prepare support for multiple Git remotes.
All tables that stores project and branch now have a new column:
remote.  This makes it possible to have a project repository that
pulls changes from more than one remote.

Specifically, add a remote column to the following database tables:

    build
    server_status

dbconvert.sql fills in "origin" in all existing entries.

Updated all code to use the remotes.

Note: because this commit does not change where things is stored in
the local filesystem, it only works if remote is set to "origin".  A
future commit will fix that.