Fix indentation and error recovery in prepare_project().
Since $fulldir is created in this function, the "cd $fulldir" that
follows should never fail.  Exit if it should fail anyhow.  (The code
used to run the NEWCHECK ls command only if the cd succeeded.  The
rest of the function was always run, though, which was probably not
intended.)

Call wget_exit immediately if wget fails, instead of chaining it with
the following if statement.