Use curl instead of put.  Remove put.
This gives us support for using https to upload the compiled sources.
That is good, since it makes it a lot harder for a man-in-the-middle
attack.  It also simplifies the client package a lot.

This patch keeps some support code for building put.  That code isn't
used right now, but a future commit will reuse it to build the new
retouch program introduced by that commit.