Work around a curl 7.21 issue that caused repeated downloads.
If using curl 7.21 or earlier, we increase the timestamp of the
downloaded snapshot.tar.gz one second so that the -z option works
properly.  If GNU stat and GNU touch are installed, they are used,
otherwise the new retouch program is compiled and used.

If we are using curl 7.22.0 or newer, we don't need to modify the
timestamp.

This reuses the old dormant machinery that were previously used to
build put.