Don't modify the database when receiving a duplicate binary package.
Rename low_process_package() to parse_package().

Call store_result() directly from process_package() instead of doing
it in low_process_package().  This removes one redundant check of the
dry_run variable.  More importantly, it lets us avoid modifying the
database if we reject a binary package because we already had a result
from that system for that source distribution.

The new name makes the responsibility of the function more clear:
parse everything we need to know about, but don't modify any state yet
(except perhaps creating a new system).  Modifying the state is the
responsibility of store_result().